+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import TodayWeather from "./screens/TodayWeather";
-import FutureWeather from "./screens/TodayWeather";
+import FutureWeather from "./screens/FutureWeather";
 import SelectPlace from "./screens/SelectPlace";
 
 const App = () => {
@@ -10,7 +10,7 @@ const App = () => {
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<SelectPlace />} />
-          <Route path="/today-weather" element={<TodayWeather />} />
+          <Route path="/today-weather/" element={<TodayWeather />} />
           <Route path="/future-weather/" element={<FutureWeather />} />
         </Routes>
       </BrowserRouter>
